worker, batch の共通化
