+++ application/batch/entity/sqs.go
@@ -1,10 +1,10 @@
 package entity
 
-type IOSVoIPPushAlarmInfoSQSMessage struct {
-	AlarmID        string `json:"alarmID"`
-	UserID         string `json:"userID"`
-	SNSEndpointArn string `json:"snsEndpointArn"`
-	CharaID        string `json:"charaID"`
-	CharaName      string `json:"charaName"`
-	VoiceFileURL   string `json:"voiceFileURL"`
-}
+//type IOSVoIPPushAlarmInfoSQSMessage struct {
+//	AlarmID        string `json:"alarmID"`
+//	UserID         string `json:"userID"`
+//	SNSEndpointArn string `json:"snsEndpointArn"`
+//	CharaID        string `json:"charaID"`
+//	CharaName      string `json:"charaName"`
+//	VoiceFileURL   string `json:"voiceFileURL"`
+//}
