+++ application/batch/service/batch.go
@@ -4,12 +4,9 @@ import (
 	"errors"
 	"fmt"
 	"github.com/takoikatakotako/charalarm/batch/database"
-	"github.com/takoikatakotako/charalarm/batch/entity"
 	"github.com/takoikatakotako/charalarm/batch/logger"
-
-	//"github.com/takoikatakotako/charalarm/batch/sqs"
-	//"github.com/takoikatakotako/charalarm/batch/logger"
-	"github.com/takoikatakotako/charalarm/batch/repository"
+	"github.com/takoikatakotako/charalarm/repository"
+	"github.com/takoikatakotako/charalarm/repository/entity"
 	"math/rand"
 	"runtime"
 	"time"
