+++ application/repository/aws_dynamodb_alarm.go
