+++ application/repository/aws_dynamodb_chara.go
