+++ application/repository/aws_dynamodb_chara_test.go
