+++ application/repository/entity/sns.go
@@ -1,0 +1,7 @@
+package entity
+
+type IOSVoIPPushSNSMessage struct {
+	CharaID      string `json:"charaID"`
+	CharaName    string `json:"charaName"`
+	VoiceFileURL string `json:"voiceFileURL"`
+}
