+++ application/repository/entity/sqs.go
@@ -1,0 +1,10 @@
+package entity
+
+type IOSVoIPPushAlarmInfoSQSMessage struct {
+	AlarmID        string `json:"alarmID"`
+	UserID         string `json:"userID"`
+	SNSEndpointArn string `json:"snsEndpointArn"`
+	CharaID        string `json:"charaID"`
+	CharaName      string `json:"charaName"`
+	VoiceFileURL   string `json:"voiceFileURL"`
+}
