+++ application/repository/environment.go
@@ -1,8 +1,8 @@
 package repository
 
-const (
-	LocalstackEndpoint = "http://localhost:4566"
-)
+//const (
+//	LocalstackEndpoint = "http://localhost:4566"
+//)
 
 type Environment struct {
 	ResourceBaseURL string
