+++ application/worker/entity/entity.go
@@ -1,16 +1,17 @@
 package entity
 
-type IOSVoIPPushAlarmInfoSQSMessage struct {
-	AlarmID        string `json:"alarmID"`
-	UserID         string `json:"userID"`
-	SNSEndpointArn string `json:"snsEndpointArn"`
-	CharaID        string `json:"charaID"`
-	CharaName      string `json:"charaName"`
-	VoiceFileURL   string `json:"voiceFileURL"`
-}
-
-type IOSVoIPPushSNSMessage struct {
-	CharaID      string `json:"charaID"`
-	CharaName    string `json:"charaName"`
-	VoiceFileURL string `json:"voiceFileURL"`
-}
+//
+//type IOSVoIPPushAlarmInfoSQSMessage struct {
+//	AlarmID        string `json:"alarmID"`
+//	UserID         string `json:"userID"`
+//	SNSEndpointArn string `json:"snsEndpointArn"`
+//	CharaID        string `json:"charaID"`
+//	CharaName      string `json:"charaName"`
+//	VoiceFileURL   string `json:"voiceFileURL"`
+//}
+//
+//type IOSVoIPPushSNSMessage struct {
+//	CharaID      string `json:"charaID"`
+//	CharaName    string `json:"charaName"`
+//	VoiceFileURL string `json:"voiceFileURL"`
+//}
